+++ src/main/java/com/unusualmodding/opposing_force/entity/Dicer.java
@@ -2,13 +2,11 @@ package com.unusualmodding.opposing_force.entity;
 
 import com.unusualmodding.opposing_force.entity.ai.goal.DicerAttackGoal;
 import com.unusualmodding.opposing_force.entity.ai.goal.DicerLaserGoal;
-import com.unusualmodding.opposing_force.entity.ai.navigation.SmoothGroundPathNavigation;
-import com.unusualmodding.opposing_force.entity.utils.AttackState;
-import com.unusualmodding.opposing_force.entity.utils.EliteVariant;
+import com.unusualmodding.opposing_force.entity.base.OPMonster;
 import com.unusualmodding.opposing_force.entity.utils.OPPoses;
 import com.unusualmodding.opposing_force.registry.OPItems;
 import com.unusualmodding.opposing_force.registry.OPSoundEvents;
-import com.unusualmodding.opposing_force.world.OPWorldData;
+import com.unusualmodding.opposing_force.utils.SmoothAnimationState;
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.syncher.EntityDataAccessor;
@@ -16,7 +14,6 @@ import net.minecraft.network.syncher.EntityDataSerializers;
 import net.minecraft.network.syncher.SynchedEntityData;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundEvents;
-import net.minecraft.util.Mth;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.Difficulty;
 import net.minecraft.world.DifficultyInstance;
@@ -30,10 +27,8 @@ import net.minecraft.world.entity.ai.goal.RandomLookAroundGoal;
 import net.minecraft.world.entity.ai.goal.WaterAvoidingRandomStrollGoal;
 import net.minecraft.world.entity.ai.goal.target.HurtByTargetGoal;
 import net.minecraft.world.entity.ai.goal.target.NearestAttackableTargetGoal;
-import net.minecraft.world.entity.ai.navigation.PathNavigation;
 import net.minecraft.world.entity.animal.IronGolem;
 import net.minecraft.world.entity.monster.Creeper;
-import net.minecraft.world.entity.monster.Monster;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
@@ -43,31 +38,24 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 @SuppressWarnings("deprecation")
-public class Dicer extends Monster implements AttackState, EliteVariant {
+public class Dicer extends OPMonster {
 
-    private static final EntityDataAccessor<Integer> ATTACK_STATE = SynchedEntityData.defineId(Dicer.class, EntityDataSerializers.INT);
-    private static final EntityDataAccessor<Boolean> RUNNING = SynchedEntityData.defineId(Dicer.class, EntityDataSerializers.BOOLEAN);
     private static final EntityDataAccessor<Boolean> LASERING = SynchedEntityData.defineId(Dicer.class, EntityDataSerializers.BOOLEAN);
-    private static final EntityDataAccessor<Boolean> ARCH_DICER = SynchedEntityData.defineId(Dicer.class, EntityDataSerializers.BOOLEAN);
 
-    public final AnimationState idleAnimationState = new AnimationState();
-    public final AnimationState slash1AnimationState = new AnimationState();
-    public final AnimationState slash2AnimationState = new AnimationState();
-    public final AnimationState crossSlashAnimationState = new AnimationState();
-    public final AnimationState laserAnimationState = new AnimationState();
-    public final AnimationState tailSpinAnimationState = new AnimationState();
+    public final SmoothAnimationState slash1AnimationState = new SmoothAnimationState(1.0F);
+    public final SmoothAnimationState slash2AnimationState = new SmoothAnimationState(1.0F);
+    public final SmoothAnimationState crossSlashAnimationState = new SmoothAnimationState();
+    public final SmoothAnimationState laserAnimationState = new SmoothAnimationState();
+    public final SmoothAnimationState tailSpinAnimationState = new SmoothAnimationState();
 
-    private int slashTicks;
-    private int crossSlashTicks;
-    private int tailSpinTicks;
-    private int laserTicks;
+    public boolean slashAlt = false;
 
     public int laserCooldown = 100 + this.getRandom().nextInt(100);
     public int slashCooldown = 0;
     public int crossSlashCooldown = 0;
     public int tailSpinCooldown = 0;
 
-    public Dicer(EntityType<? extends Monster> entityType, Level level) {
+    public Dicer(EntityType<? extends OPMonster> entityType, Level level) {
         super(entityType, level);
         this.xpReward = 10;
     }
@@ -93,11 +81,6 @@ public class Dicer extends Monster implements AttackState, EliteVariant {
     }
 
     @Override
-    protected @NotNull PathNavigation createNavigation(@NotNull Level level) {
-        return new SmoothGroundPathNavigation(this, level);
-    }
-
-    @Override
     public float getStepHeight() {
         if (this.getAttackState() == 3) {
             return 1.1F;
@@ -120,36 +103,16 @@ public class Dicer extends Monster implements AttackState, EliteVariant {
             if (tailSpinCooldown > 0) tailSpinCooldown--;
             if (crossSlashCooldown > 0) crossSlashCooldown--;
         }
-
-        if (slashTicks > 0) slashTicks--;
-        if (crossSlashTicks > 0) crossSlashTicks--;
-        if (tailSpinTicks > 0) tailSpinTicks--;
-        if (laserTicks > 0) laserTicks--;
-
-        if (slashTicks == 0 && this.getPose() == OPPoses.SLASHING.get()) this.setPose(Pose.STANDING);
-        if (crossSlashTicks == 0 && this.getPose() == OPPoses.CROSS_SLASHING.get()) this.setPose(Pose.STANDING);
-        if (tailSpinTicks == 0 && this.getPose() == OPPoses.TAIL_SPINNING.get()) this.setPose(Pose.STANDING);
-        if (laserTicks == 0 && this.getPose() == OPPoses.LASERING.get()) this.setPose(Pose.STANDING);
-
-        if (this.level().isClientSide) this.setupAnimationStates();
-    }
-
-    private void setupAnimationStates() {
-        if (slashTicks == 0 && (this.slash1AnimationState.isStarted() || this.slash2AnimationState.isStarted())) {
-            this.slash1AnimationState.stop();
-            this.slash2AnimationState.stop();
-        }
-        if (crossSlashTicks == 0 && this.crossSlashAnimationState.isStarted()) this.crossSlashAnimationState.stop();
-        if (tailSpinTicks == 0 && this.tailSpinAnimationState.isStarted()) this.tailSpinAnimationState.stop();
-        if (laserTicks == 0 && this.laserAnimationState.isStarted()) this.laserAnimationState.stop();
-        this.idleAnimationState.animateWhen(this.getPose() != OPPoses.LASERING.get() && this.getPose() != OPPoses.CROSS_SLASHING.get() && this.getPose() != OPPoses.TAIL_SPINNING.get(), this.tickCount);
     }
 
     @Override
-    public void calculateEntityAnimation(boolean flying) {
-        float f1 = (float) Mth.length(this.getX() - this.xo, this.getY() - this.yo, this.getZ() - this.zo);
-        float f2 = Math.min(f1 * 8.0F, 1.0F);
-        this.walkAnimation.update(f2, 0.4F);
+    public void setupAnimationStates() {
+        this.idleAnimationState.animateWhen(this.getPose() != OPPoses.LASERING.get() && this.getPose() != OPPoses.CROSS_SLASHING.get() && this.getPose() != OPPoses.TAIL_SPINNING.get(), this.tickCount);
+        this.slash1AnimationState.animateWhen(this.getPose() == OPPoses.ATTACKING.get() && !slashAlt, this.tickCount);
+        this.slash2AnimationState.animateWhen(this.getPose() == OPPoses.ATTACKING.get() && slashAlt, this.tickCount);
+        this.crossSlashAnimationState.animateWhen(this.getPose() == OPPoses.CROSS_SLASHING.get(), this.tickCount);
+        this.tailSpinAnimationState.animateWhen(this.getPose() == OPPoses.TAIL_SPINNING.get(), this.tickCount);
+        this.laserAnimationState.animateWhen(this.getPose() == OPPoses.LASERING.get(), this.tickCount);
     }
 
     @Override
@@ -166,81 +129,9 @@ public class Dicer extends Monster implements AttackState, EliteVariant {
     }
 
     @Override
-    public void onSyncedDataUpdated(@NotNull EntityDataAccessor<?> entityDataAccessor) {
-        if (DATA_POSE.equals(entityDataAccessor)) {
-            if (this.getPose() == OPPoses.SLASHING.get()) {
-                this.slashTicks = 20;
-                if (this.getRandom().nextBoolean()) {
-                    this.slash2AnimationState.start(this.tickCount);
-                } else {
-                    this.slash1AnimationState.start(this.tickCount);
-                }
-            }
-            else if (this.getPose() == OPPoses.CROSS_SLASHING.get()) {
-                this.crossSlashTicks = 50;
-                this.crossSlashAnimationState.start(this.tickCount);
-            }
-            else if (this.getPose() == OPPoses.TAIL_SPINNING.get()) {
-                this.tailSpinTicks = 20;
-                this.tailSpinAnimationState.start(this.tickCount);
-            }
-            else if (this.getPose() == OPPoses.LASERING.get()) {
-                this.laserTicks = 100;
-                this.laserAnimationState.start(this.tickCount);
-            }
-            else if (this.getPose() == Pose.STANDING) {
-                this.slash1AnimationState.stop();
-                this.slash2AnimationState.stop();
-                this.crossSlashAnimationState.stop();
-                this.laserAnimationState.stop();
-            }
-        }
-        super.onSyncedDataUpdated(entityDataAccessor);
-    }
-
-    @Override
     protected void defineSynchedData() {
         super.defineSynchedData();
-        this.entityData.define(ATTACK_STATE, 0);
-        this.entityData.define(RUNNING, false);
         this.entityData.define(LASERING, false);
-        this.entityData.define(ARCH_DICER, false);
-    }
-
-    @Override
-    public void addAdditionalSaveData(@NotNull CompoundTag compoundTag) {
-        super.addAdditionalSaveData(compoundTag);
-        compoundTag.putInt("AttackState", this.getAttackState());
-        compoundTag.putBoolean("Running", this.isRunning());
-        compoundTag.putBoolean("Lasering", this.isLasering());
-        compoundTag.putBoolean("ArchDicer", this.isElite());
-    }
-
-    @Override
-    public void readAdditionalSaveData(@NotNull CompoundTag compoundTag) {
-        super.readAdditionalSaveData(compoundTag);
-        this.setAttackState(compoundTag.getInt("AttackState"));
-        this.setRunning(compoundTag.getBoolean("Running"));
-        this.setLasering(compoundTag.getBoolean("Lasering"));
-        this.setElite(compoundTag.getBoolean("ArchDicer"));
-    }
-
-    public boolean isRunning() {
-        return this.entityData.get(RUNNING);
-    }
-
-    public void setRunning(boolean running) {
-        this.entityData.set(RUNNING, running);
-    }
-
-    @Override
-    public int getAttackState() {
-        return this.entityData.get(ATTACK_STATE);
-    }
-
-    @Override
-    public void setAttackState(int state) {
-        this.entityData.set(ATTACK_STATE, state);
     }
 
     public boolean isLasering() {
@@ -249,16 +140,6 @@ public class Dicer extends Monster implements AttackState, EliteVariant {
 
     public void setLasering(boolean lasering) {
         this.entityData.set(LASERING, lasering);
-    }
-
-    @Override
-    public boolean isElite() {
-        return this.entityData.get(ARCH_DICER);
-    }
-
-    @Override
-    public void setElite(boolean elite) {
-        this.entityData.set(ARCH_DICER, elite);
     }
 
     @Override