+++ Changelog.md
@@ -9,6 +9,7 @@
 - Skyvern heads now drop from the normal loot table
 - Slug eggs can now be smelted into slimeballs
 - Tarts can now be cooked on a campfire
+- All opposing force mobs now extend a base opposing force monster class
 
 # 2.0.2
 - Fixed fire slimes not dropping fire gel when spit from a guzzler